CrawlerService/requirements.txt
avro==1.12.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
confluent-kafka==2.6.1
ebook-searching-proto==0.1.3
greenlet==3.1.1
grpcio==1.68.0
idna==3.10
mysql-connector-python==9.1.0
protobuf==3.20.3
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uplink==0.9.7
uritemplate==4.1.1
urllib3==2.2.3
